Implement  mouse scroll for camera controller

+++ GUIApp/include/IController.hpp
@@ -6,5 +6,5 @@ public:
 	virtual ~IController() {}
 	virtual void resetMousePos(double posX, double posY) = 0;
 	virtual void moveMouse(double deltaX, double deltaY) = 0;
-	virtual void scrollMouse(int scrollStep) = 0;
+	virtual void scrollMouse(double scrollStep) = 0;
 };
